# Add secs2date function and speed up date2secs