## Description

By popular demand this brings fast conversion from CXC time to Year Day-of-Year date via a new function `secs2date`.
See #27 for overall motivation.

## Testing

- [x] Passes unit tests on MacOS
- [x] Functional testing

### Functional performance testing

```
>>> from cxotime import CxoTime, secs2date, date2secs
>>> secs = CxoTime('2020:001').secs

>>> %timeit CxoTime(secs, format='secs').date
377 µs ± 3.14 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

>>> %timeit secs2date(secs)
22.1 µs ± 239 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> from Chandra.Time import secs2date as s2d
>>> %timeit s2d(secs)
14.9 µs ± 245 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

The `date2secs` speed-up:
```
# NEW
>>> %timeit date2secs('2020:001')
21.9 µs ± 2.19 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

# PREVIOUS (current master)
>>> %timeit date2secs('2020:001')
38.8 µs ± 2.9 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```


Fixes #5